feat: front-end for model Produtos

+++ semi-joias-admin/sales/views/produtos_view.py
@@ -27,12 +27,15 @@ def exibir_produtos(request, id):
         'product_name_filter': product_name,
     })
     
-def cadastrar_produto(request):
+def cadastrar_produto(request, id):
+    
+    maleta = get_object_or_404(Maleta, id=id)
+    
     if request.method == "GET":
-        return render(request, 'sales/registrar_produto.html')
+        return render(request, 'sales/registrar_produto.html', {'maleta': maleta})
     
     elif request.method == "POST":
-        product_briefcase = request.POST.get('product_briefcase')
+        product_briefcase = maleta
         product_name = request.POST.get('product_name')
         product_code = request.POST.get('product_code')
         product_value = request.POST.get('product_value')
@@ -46,26 +49,34 @@ def cadastrar_produto(request):
             product_quantity=product_quantity,
         )
         produto.save()
-        return redirect('sales:info_maleta')
+        return redirect('sales:maleta_produtos', id=id)
 
 
 def atualizar_produto(request, id):
     produto = get_object_or_404(Produtos, id=id)
     
+    maleta = produto.product_briefcase
+    
     if request.method == "GET":
-        return render(request, 'sales/atualizar_produto.html', {'produto': produto})
+        contexto = {
+            'produto': produto,
+            'maleta': maleta
+        }
+        
+        return render(request, 'sales/atualizar_produto.html', contexto)
     
     elif request.method == "POST":
-        produto.product_briefcase = request.POST.get('product_briefcase')
         produto.product_name = request.POST.get('product_name')
         produto.product_code = request.POST.get('product_code')
         produto.product_value = request.POST.get('product_value')
         produto.product_quantity = request.POST.get('product_quantity')
         
         produto.save()
-        return redirect('sales:info_maleta')
+        return redirect('sales:maleta_produtos', id=id)
 
 def deletar_produto(request, id):
     produto = get_object_or_404(Produtos, id=id)
-    produto.delete()
-    return redirect('sales:info_maleta')
+    if request.method == "POST":
+        maleta_id = produto.product_briefcase.id
+        produto.delete()
+        return redirect('sales:maleta_produtos', id=maleta_id)
